🐛 [Auth] Fixed API token generation when 2FA is enabled.
It should suffice for the user to provide username+password when
creating a new API token, even if 2FA is enabled.

That's because user authentication has already occurred by the time that
that check is made, and the user is already logged through a valid
session or API token, so adding an 2FA code check isn't required.

This also ensures that the UI doesn't break with a 401 on
`/#settings?page=tokens&type=api` when creating a new token.