⚠️  Ensure that Websocket connections are always terminated upon auth failure.
In Tornado there can apparently be some race condition where `open` on a
Websocket handler does a `self.close()`, but the client is still sending
some bytes.

In that case, it may happen that the extra message is still processed.

This commit prevents the race condition by raising an exception in
`open` upon authentication failure instead of doing `close()+return`.