Support (and ignore) optional arguments on `inspect` actions.
This is namely because earlier versions supported `html_doc` (and a few
other arguments) on the `inspect.get_all_plugins` action.

That option is still used on the Platypush web extension (which hasn't
been updated in a while), and it breaks the action autocomplete form.

We should still support options for the `inspect` actions, and just
ignore them if they are not supported.