[media] Pre-emptively override YouTube URL metadata
When using e.g. the Invidious or NewPipe handlers to process videos, the
video metadata parsed from the API may include the URL to the local
instance, not the YouTube URL.

This causes the player plugin downstream to break.

We should instead ensure that, when requesting to play a resource that
matches `/watch?v=([^&]+)`, the YouTube URL is propagated down to the
player - and conversely to the yt-dlp process.