[notes] Run deduplication on the state delta before persisting it
This is to prevent cases where e.g. a draft note is deleted and a new
one is persisted under the same path (but with different IDs).

Other plugins rely on the full path of a note to infer their
synchronization state. Such cases result in state deltas that have both
added and deleted events for the same path, which can result in
undefined behaviour.

It is instead required to scan the state delta before processing it and
group changed notes by path, and then pick only the most recently
updated one.