[notes] Added `conflicting_for` property to notes
If the note is a _virtual_ note, and it's actually a copy of a remote
note that a locally synced one is in conflict with, then
`conflicting_for` will contain the local conflicting copy, while the
current note contains a copy of the remote reference.

If a note has `conflicting_for` set, it will also have
`is_conflict_note` set to True.